fix issue in unlinking log file if there is no file path

+++ app/Support/SAPCustomer.php
@@ -65,7 +65,9 @@ class SAPCustomer
             if(in_array($response->getStatusCode(), [200,201])){
                 $response = json_decode($response->getBody(),true);
 
-                unlink($logFilePath);
+                if($logFilePath){
+                    unlink($logFilePath);
+                }
 
                 return array(
                                 'status' => true,
